♻️ Move GitHubPullRequestStatus logic to a separate module
